refactor(core): extract serializer.ts from operations.ts
Move exportMemories() and importMemories() into dedicated
serializer.ts module. Re-export from operations.ts to maintain
backward compatibility with all transports.
